agent-runner/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
anthropic>=0.42.0
pydantic==2.10.3
pyyaml>=6.0
anyio>=4.0.0
